+++ hydromt_delft3dfm/components/mesh.py
@@ -12,6 +12,8 @@ import xugrid as xu
 from hydromt import hydromt_step
 from hydromt.model import Model
 from hydromt.model.components import MeshComponent
+from pyproj import CRS
+from shapely.geometry import box
 
 from hydromt_delft3dfm import workflows
 from hydromt_delft3dfm.utils import io_utils, mesh_utils
@@ -53,6 +55,43 @@ class DFlowFMMeshComponent(MeshComponent):
             model=model,
             filename=filename,
         )
+
+    # TODO: improved hydromt-core method, fix in hydromt-core instead
+    @property
+    def crs(self) -> CRS | None:
+        """Returns model mesh crs."""
+        if not self.is_empty:
+            return next(iter(self.data.ugrid.crs.values()))
+        return None
+
+    # TODO: improved hydromt-core method, fix in hydromt-core instead
+    @property
+    def bounds(self) -> tuple[float, float, float, float] | None:
+        """Returns model mesh bounds."""
+        if not self.is_empty:
+            return self.data.ugrid.bounds
+        return None
+
+    # TODO: improved hydromt-core method, fix in hydromt-core instead
+    @property
+    def _region_data(self) -> gpd.GeoDataFrame | None:
+        """Return mesh total_bounds as a geodataframe."""
+        if not self.is_empty:
+            region = gpd.GeoDataFrame(
+                geometry=[box(*self.data.ugrid.total_bounds)], crs=self.crs
+            )
+            return region
+        return None
+
+    # TODO: add to hydromt-core instead
+    @property
+    def is_empty(self):
+        """Check whether the mesh is empty or not."""
+        ndims = len(self.data.sizes)
+        if ndims == 0:
+            return True
+        else:
+            return False
 
     ### I/O methods ###
     @hydromt_step
@@ -104,6 +143,7 @@ class DFlowFMMeshComponent(MeshComponent):
     def write(self, write_gui: bool = True) -> None:
         """Write 1D branches and 2D mesh at <root/dflowfm/fm_net.nc>."""
         self.root.is_writing_mode()
+        logger.info("Writing mesh file.")
         savedir = join(self.root.path, "dflowfm")
         Path(savedir).mkdir(parents=True, exist_ok=True)
         mesh_filename = "fm_net.nc"