+++ tests/test_dflowfm.py
@@ -7,28 +7,60 @@ TESTDATADIR = join(dirname(abspath(__file__)), "data")
 EXAMPLEDIR = join(dirname(abspath(__file__)), "..", "examples")
 TOLERANCE = 1e-6
 
+
+def test_write_empty_model(tmpdir):
+    """
+    failed before: https://github.com/Deltares/hydromt_delft3dfm/issues/246
+    """
+    root = join(tmpdir, "dflowfm_example")
+    mod1 = DFlowFMModel(
+        root=root,
+        mode="w",
+        crs=3857,
+    )
+    mod1.write()
+
+
+def test_write_readonlymode(tmpdir, caplog):
+    root = join(tmpdir, "dflowfm_example")
+    mod1 = DFlowFMModel(
+        root=root,
+        mode="w",
+        crs=3857,
+    )
+    mod1.write()
+    
+    mod2 = DFlowFMModel(
+        root=root,
+        mode="r",
+        crs=3857,
+    )
+    mod2.write()
+    assert "Cannot write in read-only mode" in caplog.text
+
+
 def test_read_write_config_empty_paths(tmpdir):
     # Instantiate an empty model
-    dir_root = join(EXAMPLEDIR, "dflowfm_piave")
-    dir_model = join(tmpdir, "dflowfm_piave")
-    import shutil
-    shutil.copytree(dir_root, dir_model)
-    model = DFlowFMModel(root=dir_model, mode="r+")
+    root = join(tmpdir, "dflowfm_example")
+    model1 = DFlowFMModel(
+        root=root,
+        mode="w",
+        crs=3857,
+    )
     # Get the mdu settings
-    model.mdu.read()
+    model1.mdu.read()
     # Check whether the path is an emtpy string
-    # TODO: we temporarly put . in the example mdu, so this is now also here
-    assert model.mdu.data["output"]["outputdir"] == Path(".") 
+    assert model1.mdu.data["output"]["outputdir"] == ""
     
-    # write the mdu to read again
-    model.mdu.write()
-    # re-read the model
-    model2 = DFlowFMModel(root=dir_model, mode="r")
+    # write the model to read it again
+    model1.write()
+    model2 = DFlowFMModel(root=root, mode="r", crs=3857)
     # Get the mdu settings
     model2.mdu.read()
     # Check whether the path is an emtpy string
     # TODO: should be an empty string: https://github.com/Deltares/HYDROLIB-core/issues/703
     # then update this test: https://github.com/Deltares/hydromt_delft3dfm/issues/148
+    # and update the reference mdu files for piave and local
     assert model2.mdu.data["output"]["outputdir"] == Path(".")