+++ docs/changelog.rst
@@ -23,6 +23,8 @@ Changed
 Fixed
 -----
 - Write geoms only once (PR #224)
+- support for writing models without mesh/inifield/forcing (PR #248)
+- fixed crs/bounds/region properties of DFlowFMMeshComponent (PR #251)
 
 
 v0.3.0 (18 October 2024)